- All methods are tested

+++ src/Car.java
@@ -6,8 +6,8 @@ abstract class Car {
     protected double currentSpeed; // The current speed of the car
     protected Color color; // Color of the car
     protected String modelName; // The car model name
-    protected int x;
-    protected int y;
+    protected double x;
+    protected double y;
     protected int dir; // The Car's direction in Degrees
     protected boolean turboOn;
     protected boolean engineOn;
